Declare OptionValue Enum variants in descending order of complexity
Since Serde tries to match the JSON data with the Enum variants in order of declaration, it is possible to match on variants that do not capture all the data. To avoid this, declare the enum variants in descending order of complexity so that Serde deserializes the data with as much detail as possible.